requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0
numpy==1.26.3